Move render-block route to theme
