+++ packages/global/routes/index.js
@@ -1,11 +1,11 @@
 const htmlSitemap = require('@parameter1/base-cms-marko-web-html-sitemap/routes');
+const renderBlock = require('@randall-reilly/package-theme-monorail/routes/render-block');
 const feed = require('./feed');
 const nativeX = require('./native-x');
 const omedaNewsletters = require('./omeda-newsletters');
 const printContent = require('./print-content');
 const publicFiles = require('./public-files');
 const redirects = require('./redirects');
-const renderBlock = require('./render-block');
 const search = require('./search');
 const staticPage = require('./static-page');
 const taxonomy = require('./taxonomy');
